Rename `extract` command to `extract crd`
